feat(web): 👻 Lobby | group path
- The Lobby was separated as a self-contained folder for better scalability.
