--- web/app/[locale]/game/_ui/chat/messages/index.tsx
@@ -1,1 +1,0 @@
-export * from './Mood';
