+++ web/app/[locale]/game/_ui/index.ts
@@ -1,2 +1,1 @@
 export * from './layout';
-export * from './chat';
